requirements.txt
azure-functions
requests
pymessenger